+++ src/components/Answers/index.tsx
@@ -1,3 +1,4 @@
+import { useCallback, useMemo } from "react";
 import { useAppSelector } from "../../redux/store";
 import { shuffleArray } from "../../utils/array";
 import { MultipleAnswers } from "./Multiple";
@@ -6,10 +7,14 @@ import { SingleAnswers } from "./Single";
 export const Answers = () => {
   const { currentQuestion } = useAppSelector((state) => state.questionsSlice);
 
-  const anwers = shuffleArray([
-    ...(currentQuestion?.incorrect_answers as string[]),
-    currentQuestion?.correct_answer || "",
-  ]);
+  const anwers = useMemo(
+    () =>
+      shuffleArray([
+        ...(currentQuestion?.incorrect_answers as string[]),
+        currentQuestion?.correct_answer || "",
+      ]),
+    [currentQuestion]
+  );
 
   return (
     <div>
