+++ src/components/Questions/index.tsx
@@ -1,33 +1,12 @@
 import { Button, Form } from "antd";
-import { useAppDispatch, useAppSelector } from "../../redux/store";
-import { useEffect } from "react";
-import { fetchQuestions } from "../../redux/questions/api/asyncActions";
 import { Question } from "../Question";
-import { moveToNextQuestion } from "../../redux/questions/slice";
 import { QuestionSteps } from "../QuestionSteps";
+import { useQuestions } from "./hooks/useQuestions";
+import { useQuestionsForm } from "./hooks/useQuestionsForm";
 
 export const Questions = () => {
-  const dispatch = useAppDispatch();
-  const [form] = Form.useForm();
-
-  useEffect(() => {
-    dispatch(fetchQuestions());
-  }, []);
-
-  const nextQuestion = () => {
-    dispatch(moveToNextQuestion());
-  };
-
-  const { isLoading, currentQuestion } = useAppSelector(
-    (state) => state.questionsSlice
-  );
-
-  const onSubmit = () => {
-    form.validateFields().then((data) => {
-      console.log("data", data);
-      nextQuestion();
-    });
-  };
+  const { isLoading, currentQuestion, nextQuestion } = useQuestions();
+  const { form, isDisabled, onSubmit } = useQuestionsForm(nextQuestion);
 
   return isLoading ? (
     "Loading"
@@ -35,7 +14,7 @@ export const Questions = () => {
     <Form layout="vertical" form={form} onFinish={onSubmit}>
       <QuestionSteps />
       {currentQuestion && <Question />}{" "}
-      <Button type="primary" htmlType="submit">
+      <Button type="primary" htmlType="submit" disabled={isDisabled}>
         Next
       </Button>
     </Form>
